refactor: move `env.ts` contents to where it's used
Having an extra file for methods/vars that are unlikely to change
and are only being used in specific modules, is probably an overkill.

This is the case in `env.ts` file, where we expose code related
to the lifecylce of a `GelatoClient`.
