--- test/unit/utils/env.spec.ts
@@ -1,34 +1,0 @@
-/*!
- * @license
- * Copyright 2023 Nelson Dominguez
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { GELATO_API_KEY_VAR, loadEnvConfig } from '../../../src/utils/env';
-
-describe('utils/env', () => {
-  beforeEach(() => {
-    delete process.env[GELATO_API_KEY_VAR];
-  });
-
-  describe('#loadEnvConfig()', () => {
-    it('should throw when no apiKey is provided and "throwIfNoApiKey" is set to true', () => {
-      expect(() => loadEnvConfig(true)).toThrow();
-    });
-
-    it('should not throw when no apiKey is provided and "throwIfNoApiKey" is set to false', () => {
-      expect(() => loadEnvConfig(false)).not.toThrow();
-    });
-  });
-});
